Much better.
Now. This could lead to duplicate files. We are "overwriting" not appending and the storage is per user, so maybe not a big issue.
tested upload, back and forth, uploading and saving. All works. Now we need to tab into "refreshing" the upload list and managing errors/codes in a friendly way
@alliomeria almost done!